+++ phpmyfaq/admin/assets/src/index.js
@@ -52,11 +52,14 @@ import {
 import { handleUserList, handleUsers } from './user';
 import { handleGroups } from './group';
 import { handlePasswordStrength, handlePasswordToggle } from '../../../assets/src/utils';
-import { sidebarToggle } from './utils';
+import { handleSessionTimeout, sidebarToggle } from './utils';
 import { handleTruncateSearchTerms } from './statistics/search';
 
 document.addEventListener('DOMContentLoaded', async () => {
   'use strict';
+
+  // Session timeout
+  handleSessionTimeout();
 
   // Login
   handlePasswordToggle();