+++ phpmyfaq/admin/assets/src/utils/index.js
@@ -1,3 +1,4 @@
 export * from './notifications';
+export * from './session';
 export * from './sidebar';
 export * from './utils';