+++ phpmyfaq/admin/assets/src/utils/session.js
@@ -1,0 +1,34 @@
+import { Modal } from 'bootstrap';
+
+export const handleSessionTimeout = () => {
+  const showWarning = top.document.getElementById('pmf-show-session-warning');
+  const config = { attributes: true };
+  if (showWarning) {
+    const observer = new MutationObserver(onAttributeChange);
+    observer.observe(showWarning, config);
+    reloadCurrentPage();
+  }
+};
+
+const onAttributeChange = (mutationsList) => {
+  for (let mutation of mutationsList) {
+    if (mutation.type === 'attributes' && mutation.attributeName === 'data-value') {
+      const value = mutation.target.getAttribute('data-value');
+      toggleSessionWarnungModal(value);
+    }
+  }
+};
+const toggleSessionWarnungModal = (toggle) => {
+  const sessionWarnungModal = new Modal(top.document.getElementById('sessionWarningModal'));
+  if (toggle === 'show') {
+    sessionWarnungModal.show();
+  } else {
+    sessionWarnungModal.hide();
+  }
+};
+
+const reloadCurrentPage = () => {
+  document.getElementById('pmf-button-reload-page').addEventListener('click', () => {
+    location.reload();
+  });
+};